recipe-09/src/requirements.txt
fastapi
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-sdk
uvicorn